feat: 完善RemoteBuildResultTrigger页面，可以查看最后一次信息
