+++ src/main/java/io/jenkins/plugins/remote/result/trigger/utils/RemoteJobResultUtils.java
@@ -9,6 +9,7 @@ import hudson.model.Item;
 import io.jenkins.plugins.remote.result.trigger.RemoteJenkinsServer;
 import io.jenkins.plugins.remote.result.trigger.RemoteJobInfo;
 import io.jenkins.plugins.remote.result.trigger.exceptions.UnSuccessfulRequestStatusException;
+import io.jenkins.plugins.remote.result.trigger.model.SavedJobInfo;
 import io.jenkins.plugins.remote.result.trigger.utils.ssl.SSLSocketManager;
 import okhttp3.*;
 import org.apache.commons.io.FileUtils;
@@ -213,7 +214,7 @@ public class RemoteJobResultUtils {
             }
             // prefix with job id
             String prefix = "REMOTE_" +
-                    (StringUtils.isNotEmpty(savedJobInfo.getUid()) ? savedJobInfo.uid : savedJobInfo.remoteJobName) +
+                    (StringUtils.isNotEmpty(savedJobInfo.getUid()) ? savedJobInfo.getUid() : savedJobInfo.getRemoteJobName()) +
                     "_";
             envs.putAll(generateEnvs(prefix, savedJobInfo));
         }
@@ -304,7 +305,7 @@ public class RemoteJobResultUtils {
      * @param job Jenkins job
      * @return saved job infos
      */
-    private static List<SavedJobInfo> getSavedJobInfos(Item job) throws IOException {
+    public static List<SavedJobInfo> getSavedJobInfos(Item job) throws IOException {
         File file = getRemoteResultConfigFile(job);
         if (file.exists()) {
             ObjectMapper mapper = new ObjectMapper();
@@ -335,8 +336,8 @@ public class RemoteJobResultUtils {
     @SuppressWarnings({"rawtypes", "unchecked"})
     private static Map<String, String> generateEnvs(String prefix, SavedJobInfo savedJobInfo) {
         Map<String, String> envs = new HashMap<>();
-        if (savedJobInfo.result != null) {
-            SourceMap sourceMap = SourceMap.of(savedJobInfo.result);
+        if (savedJobInfo.getResult() != null) {
+            SourceMap sourceMap = SourceMap.of(savedJobInfo.getResult());
             // BUILD_NUMBER
             envs.put(prefix + "BUILD_NUMBER", sourceMap.stringValue("number"));
             // TIMESTAMP
@@ -379,91 +380,4 @@ public class RemoteJobResultUtils {
         return envs;
     }
 
-    /**
-     * Info
-     */
-    public static class SavedJobInfo {
-        private String remoteServer;
-        private String remoteJob;
-        @Deprecated
-        private String remoteJobName;
-        private String remoteJobUrl;
-        private String uid;
-        @Deprecated
-        private Integer triggerNumber;
-        private Integer checkedNumber;
-        private Map<String, Object> result;
-        private Map<String, Object> resultJson;
-
-        public String getRemoteServer() {
-            return remoteServer;
-        }
-
-        public void setRemoteServer(String remoteServer) {
-            this.remoteServer = remoteServer;
-        }
-
-        public String getRemoteJob() {
-            return remoteJob;
-        }
-
-        public void setRemoteJob(String remoteJob) {
-            this.remoteJob = remoteJob;
-        }
-
-        @Deprecated
-        public String getRemoteJobName() {
-            return remoteJobName;
-        }
-
-        @Deprecated
-        public void setRemoteJobName(String remoteJobName) {
-            this.remoteJobName = remoteJobName;
-        }
-
-        public String getRemoteJobUrl() {
-            return remoteJobUrl;
-        }
-
-        public void setRemoteJobUrl(String remoteJobUrl) {
-            this.remoteJobUrl = remoteJobUrl;
-        }
-
-        public String getUid() {
-            return uid;
-        }
-
-        public void setUid(String uid) {
-            this.uid = uid;
-        }
-
-        @Deprecated
-        public Integer getTriggerNumber() {
-            return triggerNumber;
-        }
-
-        public Integer getCheckedNumber() {
-            return checkedNumber;
-        }
-
-        public void setCheckedNumber(Integer checkedNumber) {
-            this.checkedNumber = checkedNumber;
-        }
-
-        public Map<String, Object> getResult() {
-            return result;
-        }
-
-        public void setResult(Map<String, Object> result) {
-            this.result = result;
-        }
-
-        public Map<String, Object> getResultJson() {
-            return resultJson;
-        }
-
-        public void setResultJson(Map<String, Object> resultJson) {
-            this.resultJson = resultJson;
-        }
-    }
 }
